requirements.txt
Django==4.2.26
pyTelegramBotAPI==4.29.1
python-dotenv==1.0.1
